♻️ refactor(core): split functions into files. event wrapper
